Rename layman.layer.filesystem.input_sld to layman.layer.filesystem.input_style
